+++ data-model-solr/src/main/java/org/mousephenotype/cda/solr/service/dto/ObservationDTOBase.java
@@ -118,13 +118,13 @@ public class ObservationDTOBase {
     protected String id;
 
     @Field(DATASOURCE_ID)
-    protected Integer dataSourceId;
+    protected Long dataSourceId;
 
     @Field(DATASOURCE_NAME)
     protected String dataSourceName;
 
     @Field(PROJECT_ID)
-    protected Integer projectId;
+    protected Long projectId;
 
     @Field(PROJECT_NAME)
     protected String projectName;
@@ -148,13 +148,13 @@ public class ObservationDTOBase {
     protected String dataType;
 
     @Field(PIPELINE_ID)
-    protected Integer pipelineId;
+    protected Long pipelineId;
 
     @Field(PROCEDURE_ID)
-    protected Integer procedureId;
+    protected Long procedureId;
 
     @Field(PARAMETER_ID)
-    protected Integer parameterId;
+    protected Long parameterId;
 
     @Field(STRAIN_ACCESSION_ID)
     protected String strainAccessionId;
@@ -178,16 +178,16 @@ public class ObservationDTOBase {
     protected String geneAccession;
 
     @Field(EXPERIMENT_ID)
-    protected Integer experimentId;
+    protected Long experimentId;
 
     @Field(PHENOTYPING_CENTER_ID)
-    protected Integer phenotypingCenterId;
+    protected Long phenotypingCenterId;
 
     @Field(PHENOTYPING_CENTER)
     protected String phenotypingCenter;
 
 	@Field(PRODUCTION_CENTER_ID)
-	protected Integer productionCenterId;
+	protected Long productionCenterId;
 
 	@Field(PRODUCTION_CENTER)
 	protected String productionCenter;
@@ -202,10 +202,10 @@ public class ObservationDTOBase {
     protected String colonyId;
 
     @Field(BIOLOGICAL_SAMPLE_ID)
-    protected Integer biologicalSampleId;
+    protected Long biologicalSampleId;
 
     @Field(BIOLOGICAL_MODEL_ID)
-    protected Integer biologicalModelId;
+    protected Long biologicalModelId;
 
     @Field(ZYGOSITY)
     protected String zygosity;
@@ -947,7 +947,7 @@ public class ObservationDTOBase {
 	/**
      * @return the pipelineId
      */
-    public Integer getPipelineId() {
+    public Long getPipelineId() {
         return pipelineId;
     }
 
@@ -955,14 +955,14 @@ public class ObservationDTOBase {
      * @param pipelineId
      *            the pipelineId to set
      */
-    public void setPipelineId(Integer pipelineId) {
+    public void setPipelineId(Long pipelineId) {
         this.pipelineId = pipelineId;
     }
 
     /**
      * @return the procedureId
      */
-    public Integer getProcedureId() {
+    public Long getProcedureId() {
         return procedureId;
     }
 
@@ -970,14 +970,14 @@ public class ObservationDTOBase {
      * @param procedureId
      *            the procedureId to set
      */
-    public void setProcedureId(Integer procedureId) {
+    public void setProcedureId(Long procedureId) {
         this.procedureId = procedureId;
     }
 
     /**
      * @return the parameterId
      */
-    public Integer getParameterId() {
+    public Long getParameterId() {
         return parameterId;
     }
 
@@ -985,7 +985,7 @@ public class ObservationDTOBase {
      * @param parameterId
      *            the parameterId to set
      */
-    public void setParameterId(Integer parameterId) {
+    public void setParameterId(Long parameterId) {
         this.parameterId = parameterId;
     }
 
@@ -1052,7 +1052,7 @@ public class ObservationDTOBase {
     /**
      * @return the experimentId
      */
-    public Integer getExperimentId() {
+    public Long getExperimentId() {
         return experimentId;
     }
 
@@ -1060,26 +1060,26 @@ public class ObservationDTOBase {
      * @param experimentId
      *            the experimentId to set
      */
-    public void setExperimentId(Integer experimentId) {
+    public void setExperimentId(Long experimentId) {
         this.experimentId = experimentId;
     }
 
     /**
      * @return the organisationId
      */
-    public Integer getPhenotypingCenterId() {
+    public Long getPhenotypingCenterId() {
         return phenotypingCenterId;
     }
 
-	public void setPhenotypingCenterId(Integer phenotypingCenterId) {
+	public void setPhenotypingCenterId(Long phenotypingCenterId) {
 		this.phenotypingCenterId = phenotypingCenterId;
 	}
 
-	public Integer getProductionCenterId() {
+	public Long getProductionCenterId() {
 		return productionCenterId;
 	}
 
-	public void setProductionCenterId(Integer productionCenterId) {
+	public void setProductionCenterId(Long productionCenterId) {
 		this.productionCenterId = productionCenterId;
 	}
 
@@ -1148,7 +1148,7 @@ public class ObservationDTOBase {
     /**
      * @return the biologicalSampleId
      */
-    public Integer getBiologicalSampleId() {
+    public Long getBiologicalSampleId() {
         return biologicalSampleId;
     }
 
@@ -1156,14 +1156,14 @@ public class ObservationDTOBase {
      * @param biologicalSampleId
      *            the biologicalSampleId to set
      */
-    public void setBiologicalSampleId(Integer biologicalSampleId) {
+    public void setBiologicalSampleId(Long biologicalSampleId) {
         this.biologicalSampleId = biologicalSampleId;
     }
 
     /**
      * @return the biologicalModelId
      */
-    public Integer getBiologicalModelId() {
+    public Long getBiologicalModelId() {
         return biologicalModelId;
     }
 
@@ -1171,7 +1171,7 @@ public class ObservationDTOBase {
      * @param biologicalModelId
      *            the biologicalModelId to set
      */
-    public void setBiologicalModelId(Integer biologicalModelId) {
+    public void setBiologicalModelId(Long biologicalModelId) {
         this.biologicalModelId = biologicalModelId;
     }
 
@@ -1388,11 +1388,11 @@ public class ObservationDTOBase {
         this.alleleAccession = alleleAccession;
     }
 
-    public Integer getDataSourceId() {
+    public Long getDataSourceId() {
         return dataSourceId;
     }
 
-    public void setDataSourceId(Integer dataSourceId) {
+    public void setDataSourceId(Long dataSourceId) {
         this.dataSourceId = dataSourceId;
     }
 
@@ -1404,7 +1404,7 @@ public class ObservationDTOBase {
         this.dataSourceName = dataSourceName;
     }
 
-    public Integer getProjectId() {
+    public Long getProjectId() {
         return projectId;
     }
 
@@ -1412,7 +1412,7 @@ public class ObservationDTOBase {
         return projectName;
     }
 
-    public void setProjectId(Integer projectId) {
+    public void setProjectId(Long projectId) {
         this.projectId = projectId;
     }
