+++ indexers/src/main/java/org/mousephenotype/cda/indexers/utils/PhisService.java
@@ -215,7 +215,7 @@ public class PhisService {
 			}
 
 			if(pImage.getExperimentGroup() != null){
-				image.setExperimentId(Integer.parseInt(pImage.getExperimentGroup()));
+				image.setExperimentId(Long.parseLong(pImage.getExperimentGroup()));
 			}
 
 			if(pImage.getZygosity() != null && pImage.getZygosity().size() >= 1){
